feat: 칭찬 공유 기능 추가 ( #2 )
1. 칭찬 공유 기능 추가
2. 공유 여부 설정 기능 추가 (기본값: 비공개)
