+++ supabase/schema.sql
@@ -3,6 +3,7 @@ create table public.users (
   id uuid references auth.users on delete cascade not null primary key,
   nickname text not null,
   avatar_url text,
+  is_public boolean default true,
   created_at timestamp with time zone default timezone('utc'::text, now()) not null
 );
